feat: add load balancer
