+++ nodemon.json
@@ -2,5 +2,5 @@
     "watch": ["src"],
     "ext": ".ts,.js",
     "ignore": [],
-    "exec": "npx ts-node ./src/main.ts"
+    "exec": "npx ts-node ./src/multi.ts"
 }
